+++ src/api/Controllers/StratigraphyV2Controller.cs
@@ -58,7 +58,7 @@ public class StratigraphyV2Controller : BoreholeControllerBase<StratigraphyV2>
 
             // Set ids of copied entities to zero. Entities with an id of zero are added as new entities to the DB.
             stratigraphy.Id = 0;
-            stratigraphy.Name += " (Clone)";
+            stratigraphy.Name = string.IsNullOrEmpty(stratigraphy.Name) ? "(Clone)" : $"{stratigraphy.Name} (Clone)";
             stratigraphy.IsPrimary = false;
 
             var entityEntry = await Context.AddAsync(stratigraphy).ConfigureAwait(false);
@@ -122,7 +122,7 @@ public class StratigraphyV2Controller : BoreholeControllerBase<StratigraphyV2>
 
             if (!await IsNameUnique(entity).ConfigureAwait(false))
             {
-                return Problem("Name must be unique");
+                return Problem(detail: "Name must be unique", type: ProblemType.UserError);
             }
 
             // If the stratigraphy to create is the first stratigraphy of a borehole,
@@ -162,7 +162,7 @@ public class StratigraphyV2Controller : BoreholeControllerBase<StratigraphyV2>
 
             if (!await IsNameUnique(entity).ConfigureAwait(false))
             {
-                return Problem("Name must be unique");
+                return Problem(detail: "Name must be unique", type: ProblemType.UserError);
             }
 
             entity.Date = entity.Date != null ? DateTime.SpecifyKind(entity.Date.Value, DateTimeKind.Utc) : null;
@@ -210,6 +210,11 @@ public class StratigraphyV2Controller : BoreholeControllerBase<StratigraphyV2>
 
     private async Task<bool> IsNameUnique(StratigraphyV2 entity)
     {
+        if (string.IsNullOrEmpty(entity.Name))
+        {
+            return true;
+        }
+
         var hasBoreholeStratigraphiesWithSameName = await Context.StratigraphiesV2
                 .AnyAsync(s => s.BoreholeId == entity.BoreholeId && s.Id != entity.Id && s.Name == entity.Name)
                 .ConfigureAwait(false);