+++ src/client/src/api/borehole.ts
@@ -1,7 +1,6 @@
 import { GridRowSelectionModel } from "@mui/x-data-grid";
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Codelist } from "../components/codelist.ts";
-import { useShowAlertOnError } from "../hooks/useShowAlertOnError.tsx";
 import { Observation } from "../pages/detail/form/hydrogeology/Observation.ts";
 import { referenceSystems } from "../pages/detail/form/location/coordinateSegmentConstants.ts";
 import { ReferenceSystemCode } from "../pages/detail/form/location/coordinateSegmentInterfaces.ts";
@@ -9,7 +8,7 @@ import { Workflow } from "../pages/detail/form/workflow/workflow.ts";
 import { Document, Photo, User, Workgroup } from "./apiInterfaces.ts";
 import { BoreholeGeometry } from "./boreholeGeometry.ts";
 import { Completion } from "./completion.ts";
-import { download, fetchApiV2, upload } from "./fetchApiV2.ts";
+import { download, fetchApiV2Legacy, fetchApiV2WithApiError, upload } from "./fetchApiV2.ts";
 import { BoreholeFile } from "./file/fileInterfaces.ts";
 import { Section } from "./section.ts";
 import { Stratigraphy, StratigraphyLegacy } from "./stratigraphy.ts";
@@ -91,7 +90,7 @@ export interface BoreholeV2 {
 const getIdQuery = (ids: number[] | GridRowSelectionModel) => ids.map(id => `ids=${id}`).join("&");
 
 export const exportJsonBoreholes = async (boreholeIds: number[] | GridRowSelectionModel) => {
-  return await fetchApiV2(`export/json?${getIdQuery(boreholeIds)}`, "GET");
+  return await fetchApiV2Legacy(`export/json?${getIdQuery(boreholeIds)}`, "GET");
 };
 
 export const importBoreholesCsv = async (workgroupId: number | null, combinedFormData: FormData) => {
@@ -107,77 +106,71 @@ export const importBoreholesZip = async (workgroupId: number | null, combinedFor
 };
 
 export const createBorehole = async (workgroupId: number): Promise<BoreholeV2> => {
-  return await fetchApiV2(`borehole`, "POST", { workgroupId, originalReferenceSystem: referenceSystems.LV95.code });
+  return await fetchApiV2WithApiError(`borehole`, "POST", {
+    workgroupId,
+    originalReferenceSystem: referenceSystems.LV95.code,
+  });
 };
 
 export const copyBorehole = async (boreholeId: GridRowSelectionModel, workgroupId: number | null) => {
-  return await fetchApiV2(`borehole/copy?id=${boreholeId}&workgroupId=${workgroupId}`, "POST");
+  return await fetchApiV2Legacy(`borehole/copy?id=${boreholeId}&workgroupId=${workgroupId}`, "POST");
 };
 
 export const exportCSVBorehole = async (boreholeIds: GridRowSelectionModel) => {
-  return await fetchApiV2(`export/csv?${getIdQuery(boreholeIds)}`, "GET");
+  return await fetchApiV2Legacy(`export/csv?${getIdQuery(boreholeIds)}`, "GET");
 };
 
 export const exportJsonWithAttachmentsBorehole = async (boreholeIds: number[] | GridRowSelectionModel) => {
   return await download(`export/zip?${getIdQuery(boreholeIds)}`);
 };
 
-export const fetchBoreholeById = async (id: number) => await fetchApiV2(`borehole/${id}`, "GET");
+export const fetchBoreholeById = async (id: number) => await fetchApiV2WithApiError(`borehole/${id}`, "GET");
 
 export const updateBorehole = async (borehole: BoreholeV2) => {
-  return await fetchApiV2("borehole", "PUT", borehole);
+  return await fetchApiV2WithApiError("borehole", "PUT", borehole);
 };
-export const deleteBorehole = async (id: number) => await fetchApiV2(`borehole?id=${id}`, "DELETE");
+export const deleteBorehole = async (id: number) => await fetchApiV2WithApiError(`borehole?id=${id}`, "DELETE");
 
 export const canUserEditBorehole = async (id: number) =>
-  await fetchApiV2(`permissions/canedit?boreholeId=${id}`, "GET");
+  await fetchApiV2WithApiError(`permissions/canedit?boreholeId=${id}`, "GET");
 
 export const canUserUpdateBoreholeStatus = async (id: number) =>
-  await fetchApiV2(`permissions/canchangestatus?boreholeId=${id}`, "GET");
+  await fetchApiV2WithApiError(`permissions/canchangestatus?boreholeId=${id}`, "GET");
 
 export const boreholeQueryKey = "boreholes";
 
 export const useBorehole = (id: number) => {
-  const query = useQuery({
+  return useQuery({
     queryKey: [boreholeQueryKey, id],
     queryFn: async () => {
       return await fetchBoreholeById(id);
     },
     enabled: !!id,
   });
-
-  useShowAlertOnError(query.isError, query.error);
-  return query;
 };
 
 export const canEditQueryKey = "canEditBorehole";
 export const useBoreholeEditable = (id: number) => {
   const { data: currentUser } = useCurrentUser();
-  const query = useQuery({
+  return useQuery({
     queryKey: [canEditQueryKey, currentUser?.id, id],
     queryFn: async () => {
       return await canUserEditBorehole(id);
     },
     enabled: !!id,
   });
-
-  useShowAlertOnError(query.isError, query.error);
-  return query;
 };
 
 export const canUpdateStatusQueryKey = "canUpdateBoreholeStatus";
 export const useBoreholeStatusEditable = (id: number) => {
   const { data: currentUser } = useCurrentUser();
-  const query = useQuery({
+  return useQuery({
     queryKey: [canUpdateStatusQueryKey, currentUser?.id, id],
     queryFn: async () => {
       return await canUserUpdateBoreholeStatus(id);
     },
     enabled: !!id,
   });
-
-  useShowAlertOnError(query.isError, query.error);
-  return query;
 };
 
 export const useBoreholeMutations = () => {
@@ -215,10 +208,6 @@ export const useBoreholeMutations = () => {
       });
     },
   });
-
-  useShowAlertOnError(useAddBorehole.isError, useAddBorehole.error);
-  useShowAlertOnError(useUpdateBorehole.isError, useUpdateBorehole.error);
-  useShowAlertOnError(useDeleteBorehole.isError, useDeleteBorehole.error);
   return {
     add: useAddBorehole,
     update: useUpdateBorehole,