+++ src/client/src/api/file/file.ts
@@ -7,7 +7,7 @@ import {
 } from "../../pages/detail/labeling/labelingInterfaces.tsx";
 import { ApiError, BoreholeAttachment } from "../apiInterfaces.ts";
 import { fetchCreatePngs, fetchExtractData, fetchExtractStratigraphy, fetchPageBoundingBoxes } from "../dataextraction";
-import { download, fetchApiV2, fetchApiV2Base, upload } from "../fetchApiV2.ts";
+import { download, fetchApiV2Base, fetchApiV2Legacy, fetchApiV2WithApiError, upload } from "../fetchApiV2.ts";
 import { processFileWithOCR } from "../ocr.ts";
 import { BaseLayer, LithologicalDescription } from "../stratigraphy.ts";
 import { BoreholeFile, DataExtractionResponse, maxFileSizeKB } from "./fileInterfaces.ts";
@@ -34,11 +34,11 @@ export async function uploadFile(boreholeId: number, file: File) {
 }
 
 export const detachFile = async (boreholeFileId: number) => {
-  return await fetchApiV2(`boreholefile/detachFile?boreholeFileId=${boreholeFileId}`, "POST");
+  return await fetchApiV2Legacy(`boreholefile/detachFile?boreholeFileId=${boreholeFileId}`, "POST");
 };
 
 export async function getFiles<FileResponse>(boreholeId: number): Promise<FileResponse[]> {
-  const response = await fetchApiV2(`boreholefile/getAllForBorehole?boreholeId=${boreholeId}`, "GET");
+  const response = await fetchApiV2Legacy(`boreholefile/getAllForBorehole?boreholeId=${boreholeId}`, "GET");
   if (response) {
     return response as FileResponse[];
   } else {
@@ -56,14 +56,18 @@ export const updateFile = async (
   description: string,
   isPublic: boolean,
 ) => {
-  return await fetchApiV2(`boreholefile/update?boreholeId=${boreholeId}&boreholeFileId=${boreholeFileId}`, "PUT", {
-    description: description,
-    public: isPublic,
-  });
+  return await fetchApiV2WithApiError(
+    `boreholefile/update?boreholeId=${boreholeId}&boreholeFileId=${boreholeFileId}`,
+    "PUT",
+    {
+      description: description,
+      public: isPublic,
+    },
+  );
 };
 
 export async function getDataExtractionFileInfo(boreholeFileId: number, index = 1): Promise<DataExtractionResponse> {
-  let response = await fetchApiV2(
+  let response = await fetchApiV2Legacy(
     `boreholefile/getDataExtractionFileInfo?boreholeFileId=${boreholeFileId}&index=${index}`,
     "GET",
   );