+++ src/client/src/api/stratigraphy.ts
@@ -1,10 +1,9 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import { Codelist } from "../components/codelist.ts";
 import { useResetTabStatus } from "../hooks/useResetTabStatus.ts";
-import { useShowAlertOnError } from "../hooks/useShowAlertOnError.tsx";
 import { User } from "./apiInterfaces.ts";
 import { boreholeQueryKey, BoreholeV2 } from "./borehole.ts";
-import { fetchApiV2, fetchApiV2WithApiError } from "./fetchApiV2.ts";
+import { fetchApiV2Legacy, fetchApiV2WithApiError } from "./fetchApiV2.ts";
 
 export interface StratigraphyLegacy {
   id: number;
@@ -30,7 +29,7 @@ export interface Stratigraphy {
   id: number;
   boreholeId: number;
   borehole: BoreholeV2 | null;
-  name: string;
+  name: string | null;
   date: string | null;
   isPrimary: boolean;
   created: Date | string | null;
@@ -92,69 +91,69 @@ export interface FaciesDescription extends BaseLayer {
 }
 
 // layers
-export const fetchLayerById = async (id: number): Promise<Layer> => await fetchApiV2(`layer/${id}`, "GET");
+export const fetchLayerById = async (id: number): Promise<Layer> => await fetchApiV2Legacy(`layer/${id}`, "GET");
 
 export const fetchLayersByProfileId = async (profileId: number): Promise<Layer[]> =>
-  await fetchApiV2(`layer?profileId=${profileId}`, "GET");
+  await fetchApiV2WithApiError(`layer?profileId=${profileId}`, "GET");
 
 export const updateLayer = async (layer: Layer): Promise<Layer> => {
   // remove derived objects
   delete layer.createdBy;
   delete layer.updatedBy;
-  return await fetchApiV2("layer", "PUT", layer);
+  return await fetchApiV2Legacy("layer", "PUT", layer);
 };
 
 // lithological descriptions
 export const fetchLithologicalDescriptionsByProfileId = async (
   profileId: number,
 ): Promise<LithologicalDescription[]> => {
-  return await fetchApiV2(`lithologicaldescription?stratigraphyId=${profileId}`, "GET");
+  return await fetchApiV2WithApiError(`lithologicaldescription?stratigraphyId=${profileId}`, "GET");
 };
 
 export const addLithologicalDescription = async (
   lithologicalDescription: LithologicalDescription,
 ): Promise<LithologicalDescription> => {
-  return await fetchApiV2("lithologicaldescription", "POST", lithologicalDescription);
+  return await fetchApiV2WithApiError("lithologicaldescription", "POST", lithologicalDescription);
 };
 
 export const updateLithologicalDescription = async (
   lithologicalDescription: LithologicalDescription,
 ): Promise<LithologicalDescription> => {
-  return await fetchApiV2("lithologicaldescription", "PUT", lithologicalDescription);
+  return await fetchApiV2WithApiError("lithologicaldescription", "PUT", lithologicalDescription);
 };
 
 export const deleteLithologicalDescription = async (id: number): Promise<void> => {
-  return await fetchApiV2(`lithologicaldescription?id=${id}`, "DELETE");
+  return await fetchApiV2WithApiError(`lithologicaldescription?id=${id}`, "DELETE");
 };
 
 // facies descriptions
 export const fetchFaciesDescriptionsByProfileId = async (profileId: number): Promise<FaciesDescription[]> => {
-  return await fetchApiV2(`faciesdescription?stratigraphyId=${profileId}`, "GET");
+  return await fetchApiV2WithApiError(`faciesdescription?stratigraphyId=${profileId}`, "GET");
 };
 
 export const addFaciesDescription = async (faciesDescription: FaciesDescription): Promise<FaciesDescription> => {
-  return await fetchApiV2("faciesdescription", "POST", faciesDescription);
+  return await fetchApiV2WithApiError("faciesdescription", "POST", faciesDescription);
 };
 
 export const updateFaciesDescription = async (faciesDescription: FaciesDescription): Promise<FaciesDescription> => {
-  return await fetchApiV2("faciesdescription", "PUT", faciesDescription);
+  return await fetchApiV2WithApiError("faciesdescription", "PUT", faciesDescription);
 };
 
 export const deleteFaciesDescription = async (id: number): Promise<void> => {
-  return await fetchApiV2(`faciesdescription?id=${id}`, "DELETE");
+  return await fetchApiV2WithApiError(`faciesdescription?id=${id}`, "DELETE");
 };
 
 // stratigraphy
 const fetchStratigraphiesByBoreholeId = async (boreholeId: number): Promise<StratigraphyLegacy[]> => {
-  return await fetchApiV2(`stratigraphy?boreholeId=${boreholeId}`, "GET");
+  return await fetchApiV2WithApiError(`stratigraphy?boreholeId=${boreholeId}`, "GET");
 };
 
 const createStratigraphy = async (boreholeId: number): Promise<StratigraphyLegacy> => {
-  return await fetchApiV2("stratigraphy", "POST", { boreholeId });
+  return await fetchApiV2WithApiError("stratigraphy", "POST", { boreholeId });
 };
 
 const copyStratigraphy = async (stratigraphy: StratigraphyLegacy): Promise<number> => {
-  return await fetchApiV2(`stratigraphy/copy?id=${stratigraphy.id}`, "POST");
+  return await fetchApiV2WithApiError(`stratigraphy/copy?id=${stratigraphy.id}`, "POST");
 };
 
 const updateStratigraphy = async (stratigraphy: StratigraphyLegacy): Promise<StratigraphyLegacy> => {
@@ -162,15 +161,15 @@ const updateStratigraphy = async (stratigraphy: StratigraphyLegacy): Promise<Str
   delete stratigraphy.createdBy;
   delete stratigraphy.updatedBy;
 
-  return await fetchApiV2("stratigraphy", "PUT", stratigraphy);
+  return await fetchApiV2WithApiError("stratigraphy", "PUT", stratigraphy);
 };
 
 const deleteStratigraphy = async (id: number): Promise<void> => {
-  return await fetchApiV2(`stratigraphy?id=${id}`, "DELETE");
+  return await fetchApiV2WithApiError(`stratigraphy?id=${id}`, "DELETE");
 };
 
 export const addBedrock = async (id: number): Promise<void> => {
-  return await fetchApiV2(`stratigraphy/addbedrock?id=${id}`, "POST");
+  return await fetchApiV2Legacy(`stratigraphy/addbedrock?id=${id}`, "POST");
 };
 
 export const stratigraphiesByBoreholeIdQueryKey = "stratigraphiesByBoreholeId";
@@ -199,7 +198,7 @@ export const useStratigraphiesByBoreholeId = (boreholeId?: number) =>
   useQuery({
     queryKey: [stratigraphiesByBoreholeIdQueryKey, boreholeId],
     queryFn: async () => {
-      return await fetchApiV2(`${stratigraphyController}?boreholeId=${boreholeId!}`, "GET");
+      return await fetchApiV2WithApiError(`${stratigraphyController}?boreholeId=${boreholeId!}`, "GET");
     },
     enabled: !!boreholeId,
   });
@@ -252,9 +251,6 @@ export const useStratigraphyMutations = () => {
     },
   });
 
-  useShowAlertOnError(useCopyStratigraphy.isError, useCopyStratigraphy.error);
-  useShowAlertOnError(useDeleteStratigraphy.isError, useDeleteStratigraphy.error);
-
   return {
     add: useAddStratigraphy,
     copy: useCopyStratigraphy,
@@ -306,11 +302,6 @@ export const useLegacyStratigraphyMutations = () => {
       invalidateStratigraphyQueries(queryClient, stratigraphy.boreholeId, true);
     },
   });
-
-  useShowAlertOnError(useAddStratigraphy.isError, useAddStratigraphy.error);
-  useShowAlertOnError(useCopyStratigraphy.isError, useCopyStratigraphy.error);
-  useShowAlertOnError(useUpdateStratigraphy.isError, useUpdateStratigraphy.error);
-  useShowAlertOnError(useDeleteStratigraphy.isError, useDeleteStratigraphy.error);
 
   return {
     add: useAddStratigraphy,
@@ -449,7 +440,7 @@ export const useChronostratigraphies = (stratigraphyID?: number) =>
   useQuery({
     queryKey: [chronostratigraphiesQueryKey, stratigraphyID],
     queryFn: async () => {
-      return await fetchApiV2(`chronostratigraphy?stratigraphyId=${stratigraphyID}`, "GET");
+      return await fetchApiV2WithApiError(`chronostratigraphy?stratigraphyId=${stratigraphyID}`, "GET");
     },
     enabled: !!stratigraphyID,
   });
@@ -459,7 +450,7 @@ export const useChronostratigraphyMutations = () => {
   const resetTabStatus = useResetTabStatus(["chronostratigraphy"]);
   const useAddChronostratigraphy = useMutation({
     mutationFn: async (chronostratigraphy: Chronostratigraphy) => {
-      return await fetchApiV2("chronostratigraphy", "POST", chronostratigraphy);
+      return await fetchApiV2WithApiError("chronostratigraphy", "POST", chronostratigraphy);
     },
     onSuccess: () => {
       resetTabStatus();
@@ -470,7 +461,7 @@ export const useChronostratigraphyMutations = () => {
   });
   const useUpdateChronostratigraphy = useMutation({
     mutationFn: async (chronostratigraphy: Chronostratigraphy) => {
-      return await fetchApiV2("chronostratigraphy", "PUT", chronostratigraphy);
+      return await fetchApiV2WithApiError("chronostratigraphy", "PUT", chronostratigraphy);
     },
     onSuccess: () => {
       resetTabStatus();
@@ -481,7 +472,7 @@ export const useChronostratigraphyMutations = () => {
   });
   const useDeleteChronostratigraphy = useMutation({
     mutationFn: async (chronostratigraphyId: number) => {
-      return await fetchApiV2(`chronostratigraphy?id=${chronostratigraphyId}`, "DELETE");
+      return await fetchApiV2WithApiError(`chronostratigraphy?id=${chronostratigraphyId}`, "DELETE");
     },
     onSuccess: () => {
       resetTabStatus();
@@ -504,7 +495,7 @@ export const useLithostratigraphies = (stratigraphyID?: number) =>
   useQuery({
     queryKey: [lithostratigraphiesQueryKey, stratigraphyID],
     queryFn: async () => {
-      return await fetchApiV2(`lithostratigraphy?stratigraphyId=${stratigraphyID}`, "GET");
+      return await fetchApiV2WithApiError(`lithostratigraphy?stratigraphyId=${stratigraphyID}`, "GET");
     },
     enabled: !!stratigraphyID,
   });
@@ -514,7 +505,7 @@ export const useLithostratigraphyMutations = () => {
   const resetTabStatus = useResetTabStatus(["lithostratigraphy"]);
   const useAddLithostratigraphy = useMutation({
     mutationFn: async (lithostratigraphy: Lithostratigraphy) => {
-      return await fetchApiV2("lithostratigraphy", "POST", lithostratigraphy);
+      return await fetchApiV2WithApiError("lithostratigraphy", "POST", lithostratigraphy);
     },
     onSuccess: () => {
       resetTabStatus();
@@ -525,7 +516,7 @@ export const useLithostratigraphyMutations = () => {
   });
   const useUpdateLithostratigraphy = useMutation({
     mutationFn: async (lithostratigraphy: Lithostratigraphy) => {
-      return await fetchApiV2("lithostratigraphy", "PUT", lithostratigraphy);
+      return await fetchApiV2WithApiError("lithostratigraphy", "PUT", lithostratigraphy);
     },
     onSuccess: () => {
       resetTabStatus();
@@ -536,7 +527,7 @@ export const useLithostratigraphyMutations = () => {
   });
   const useDeleteLithostratigraphy = useMutation({
     mutationFn: async (lithostratigraphyId: number) => {
-      return await fetchApiV2(`lithostratigraphy?id=${lithostratigraphyId}`, "DELETE");
+      return await fetchApiV2WithApiError(`lithostratigraphy?id=${lithostratigraphyId}`, "DELETE");
     },
     onSuccess: () => {
       resetTabStatus();